+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- `overrides` property supports manually adding dependencies (experimental for now)
 - fix `remove` command to not delete dependencies files from the scope as they might belong to other components
 - `overrides` property supports ignoring dependencies (experimental for now)
 - fix symlink to binary (or unsupported) files dependencies when installed via npm